move Updating to sim_control
